+++ imports/api/badges/definitions/_trophyHunter.js
@@ -20,7 +20,7 @@ const Description = trophyHunter => (
   <Fragment>
     This summoner is a Trophy Hunter!
     <br />
-    {trophyHunter.seasonRank ? `${trophyHunter.seasonRank}. Rank` : 'Unranked'}{' '}
+    {trophyHunter.s9Rank ? `${trophyHunter.s9Rank}. Rank` : 'Unranked'}{' '}
     {trophyHunter.rank && `(${trophyHunter.rank}. All-Time)`}
     <br />
     {trophyHunter.lastLogin && `Last Login ${moment(trophyHunter.lastLogin).calendar()}`}