+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Change Log
 
+### Added
+- Trophy Hunter Season 2019 started!
 ### Changed
 - Updated Riot API from v3 to v4 which adds a security layer for the privacy of players.
 - Updated to Riot API version 9.1.1.